rename berth icons
