+++ src/common/icons/IconApplications.tsx
