+++ src/common/icons/IconContracts.tsx
