+++ src/common/icons/IconCustomers.tsx
