+++ src/common/icons/IconDollyEmpty.tsx
