+++ src/common/icons/IconFence.tsx
