+++ src/common/icons/IconIndividual.tsx
