+++ src/common/icons/IconInvoice.tsx
