+++ src/common/icons/IconPencil.tsx
