+++ src/common/icons/IconPole.tsx
