+++ src/common/icons/IconRegisteredBoat.tsx
