+++ src/common/icons/IconSettings.tsx
