+++ src/common/icons/IconStreetLight.tsx
