+++ src/common/icons/IconTrash.tsx
