+++ src/common/icons/IconWinter.tsx
