--- src/common/icons/SettingsIconIcon.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-
-type Props = { className?: string };
-
-export default ({ className = '' }: Props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="100%"
-    height="100%"
-    viewBox="0 0 199.404 199.404"
-    className={className}
-  >
-    <path
-      d="M11 15.8A4.812 4.812 0 016.2 11c0-2.642 2.158-4.8 4.8-4.8s4.8 2.158 4.8 4.8-2.158 4.8-4.8 4.8zm7.759-11.443l-1.143-1.144-2.285.763-.307-.17a8.505 8.505 0 00-1.86-.77l-.35-.095L11.773.8h-1.588l-1.07 2.142-.344.094a7.66 7.66 0 00-1.813.762l-.312.181-2.296-.766L3.21 4.357l.762 2.29-.168.306a8.534 8.534 0 00-.77 1.863l-.095.35L.8 10.206V11.8l2.14 1.071.093.345c.171.629.431 1.245.76 1.816l.181.312-.765 2.299 1.142 1.144 2.286-.763.307.17a8.503 8.503 0 001.86.77l.345.094 1.07 2.142h1.595l1.07-2.142.344-.094a7.659 7.659 0 001.813-.762l.312-.181 2.296.766 1.142-1.144-.762-2.29.168-.306c.343-.624.6-1.243.77-1.863l.094-.345 2.139-1.071v-1.59l-2.17-1.041-.096-.353a7.685 7.685 0 00-.76-1.816l-.18-.312.765-2.299z"
-      stroke="#000"
-      stroke-width="1.6"
-      fill="none"
-      fill-rule="evenodd"
-    />
-  </svg>
-);
