+++ src/common/icons/index.ts
@@ -1,29 +1,29 @@
-export { default as AppolcationsIcon } from './ApplicationsIcon';
-export { default as BoatIcon } from './BoatIcon';
-export { default as BusinessIcon } from './BusinessIcon';
-export { default as CommunicationsIcon } from './CommunicationsIcon';
-export { default as ContractsIcon } from './ContractsIcon';
-export { default as DividedIcon } from './DividedIcon';
-export { default as DollyEmptyIcon } from './DollyEmptyIcon';
-export { default as ExclamationCircleIcon } from './ExclamationCircleIcon';
-export { default as FenceIcon } from './FenceIcon';
-export { default as GlobeIcon } from './GlobeIcon';
-export { default as IndividualIcon } from './IndividualIcon';
-export { default as InvoiceIconIcon } from './InvoiceIconIcon';
-export { default as NoBoatIcon } from './NoBoatIcon';
-export { default as PencilIcon } from './PencilIcon';
-export { default as PoleIcon } from './PoleIcon';
-export { default as RegisteredBoatIcon } from './RegisteredBoatIcon';
-export { default as RemindersIcon } from './RemindersIcon';
-export { default as SettingsIcon } from './SettingsIcon';
-export { default as StreetLightIcon } from './StreetLightIcon';
-export { default as TimesIcon } from './TimesIcon';
-export { default as TrashIcon } from './TrashIcon';
-export { default as TrestleIcon } from './TrestleIcon';
-export { default as UnregisteredBoatIcon } from './UnregisteredBoatIcon';
-export { default as UserIcon } from './UserIcon';
-export { default as WaterTapIcon } from './WaterTapIcon';
-export { default as WinterIcon } from './WinterIcon';
+export { default as IconApplications } from './IconApplications';
+export { default as IconBoat } from './IconBoat';
+export { default as IconBusiness } from './IconBusiness';
+export { default as IconCommunications } from './IconCommunications';
+export { default as IconContracts } from './IconContracts';
+export { default as IconDivided } from './IconDivided';
+export { default as IconDollyEmpty } from './IconDollyEmpty';
+export { default as IconExclamationCircle } from './IconExclamationCircle';
+export { default as IconFence } from './IconFence';
+export { default as IconGlobe } from './IconGlobe';
+export { default as IndividualIcon } from './IconIndividual';
+export { default as IconInvoice } from './IconInvoice';
+export { default as IconNoBoat } from './IconNoBoat';
+export { default as IconPencil } from './IconPencil';
+export { default as IconPole } from './IconPole';
+export { default as IconRegisteredBoat } from './IconRegisteredBoat';
+export { default as IconReminders } from './IconReminders';
+export { default as IconSettings } from './IconSettings';
+export { default as IconStreetLight } from './IconStreetLight';
+export { default as IconTimes } from './IconTimes';
+export { default as IconTrash } from './IconTrash';
+export { default as IconTrestle } from './IconTrestle';
+export { default as IconUnregisteredBoat } from './IconUnregisteredBoat';
+export { default as IconUser } from './IconUser';
+export { default as IconWaterTap } from './IconWaterTap';
+export { default as IconWinter } from './IconWinter';
 export { default as IconAngleDown } from './IconAngleDown';
 export { default as IconAngleLeft } from './IconAngleLeft';
 export { default as IconAngleRight } from './IconAngleRight';
